+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Currency/BuyBait.cs
@@ -17,7 +17,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Currency
         [Summary("Purchases bait from the baitshop. Each bait costs " +
                  "75 points (unless you have previously fished, in which it costs " +
                  "more based on your fishing level). At least one bait is required for `fishing`. " +
-                 "Supporters get 25% off!")]
+                 "Premium users get 25% off!")]
         [Remarks("<amount>")]
         public async Task Command(int amount)
         {