+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Currency/FishGame.cs
@@ -92,7 +92,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Currency
             var bonuses = new FishHandler.FishLevelBonuses(user.FishExp);
             roll *= 1 - (bonuses.BonusLuckPercent / 100);
 
-            if (premium || server.IsPremium)
+            if (premium)
                 roll *= 0.95;
             
             var fishType = GetFishType(roll);