+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Help/Ping.cs
@@ -7,7 +7,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Help
     public class Ping : KaguyaBase
     {
         [HelpCommand]
-        [Command("Ping")]
+        [Command("Ping", RunMode = RunMode.Async)]
         [Summary("Sends a ping to Discord and replies with the estimated " +
                  "round-trip latency, in milliseconds, to the gateway server.")]
         [Remarks("")]