+++ KaguyaProjectV2/KaguyaBot/Core/Program.cs
@@ -191,6 +191,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core
         {
             if (!allShardsLoggedIn) return;
 
+            await RateLimitService.Initialize();
 #if !DEBUG
             CachedPopularCommandTimer.Initialize();
             await AntiRaidService.Initialize();
@@ -198,7 +199,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core
             await StatsUpdater.Initialize();
             await KaguyaStatsLogger.Initialize();
             await AutoUnmuteHandler.Initialize();
-            await RateLimitService.Initialize();
             await RemindService.Initialize();
             await NSFWImageHandler.Initialize();
             await KaguyaPremiumExpirationHandler.Initialize();